real -> numeric in make_tables

+++ make_tables.sql
@@ -16,14 +16,14 @@ CREATE TABLE beds (
     record_id SERIAL PRIMARY KEY,
     hospital_pk VARCHAR(255) NOT NULL,
     collection_week DATE NOT NULL,
-    all_adult_hospital_beds_7_day_avg REAL,
-    all_pediatric_inpatient_beds_7_day_avg REAL,
-    all_adult_hospital_inpatient_bed_occupied_7_day_coverage REAL,
-    all_pediatric_inpatient_bed_occupied_7_day_avg REAL,
-    total_icu_beds_7_day_avg REAL,
-    icu_beds_used_7_day_avg REAL,
-    inpatient_beds_used_covid_7_day_avg REAL,
-    staffed_icu_adult_patients_confirmed_covid_7_day_avg REAL,
+    all_adult_hospital_beds_7_day_avg NUMERIC,
+    all_pediatric_inpatient_beds_7_day_avg NUMERIC,
+    all_adult_hospital_inpatient_bed_occupied_7_day_coverage NUMERIC,
+    all_pediatric_inpatient_bed_occupied_7_day_avg NUMERIC,
+    total_icu_beds_7_day_avg NUMERIC,
+    icu_beds_used_7_day_avg NUMERIC,
+    inpatient_beds_used_covid_7_day_avg NUMERIC,
+    staffed_icu_adult_patients_confirmed_covid_7_day_avg NUMERIC,
     FOREIGN KEY (hospital_pk) REFERENCES hospital(hospital_pk)
 );
